% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB_filter.R
\name{BB_filter}
\alias{BB_filter}
\title{Filters images of brilliant blue dye tracer experiments of snow}
\usage{
BB_filter(rgb_image)
}
\arguments{
\item{rgb_image}{is an array with RGB values}
}
\value{
A matrix with applied threshold where 0 is white, 1 is light blue, 2 is medium blue and 3 is dark blue
}
\description{
This function uses RGB thresholds to assign one of four colors to each pixel (white, dark blue, medium blue or light blue). Grey shadows are assigned white. The filter aims to make processing easier. It only works for images of snow.
}
\examples{
## To process raw image
raw_image <- readJPEG('path to image')
## Apply filter to raw image data
BB_filtered_image <- BB_filter(rgb_image = raw_image)
}
\author{
Johanna Bloecher, \email{bloecher@fzp.czu.cz}
}
