% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB_hor_distribution.R
\name{BB_hor_distribution}
\alias{BB_hor_distribution}
\title{Horizontal distribution of Brilliant Blue in Snow image}
\usage{
BB_hor_distribution(filtered_image, blue_selection = "all")
}
\arguments{
\item{filtered_image}{the returned matrix of the function BB_filter with values 0, 1, 2 and 3}

\item{blue_selection}{the blue to consider. By default, the value for blue_selection is 'all'
\itemize{
\item 'all' (all blues)
\item 'dark blue' (only dark blue),
\item 'medium blue' (only medium blues) or
\item 'light blue' (only light blues).
}}
}
\value{
Returns the vertical distribution of blue color as a vector.
}
\description{
A simple function that processes the result of BB_filter and returns a vector with fractions over the length of an image of a sprinkling experiment with Brilliant Blue.
}
\examples{
library(jpeg)
raw_image <- <- readJPEG(system.file("example.jpg", package = "bluesnow"))
## Apply filter to raw image data
BB_filtered_image <- BB_filter(rgb_image = raw_image)
BB_h_dist <- BB_hor_distribution(filtered_image = BB_filtered_image, blue_selection = 'all')
plot(BB_h_dist, xlab = "Width [Pixels]", ylab = "Fraction of Blue [-]")
}
\author{
Johanna Bloecher, \email{bloecher@fzp.czu.cz}
}
