% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB_distribution.R
\name{BB_fraction}
\alias{BB_fraction}
\title{Fractions of dark, medium, light and total blue}
\usage{
BB_fraction(filtered_image)
}
\arguments{
\item{filtered_image}{the returned matrix of the function BB_filter with values 0, 1, 2 and 3}
}
\value{
The fraction of total, dark, medium and light blue
}
\description{
A simple function that reads that takes the ouput of BB_filter and calculates the proportion or fraction of total blue pixels, dark blue pixels, medium blue pixels and light blue pixels.
}
\examples{
library(jpeg)
raw_image <- <- readJPEG(system.file("example.jpg", package = "bluesnow"))
## Apply filter to raw image data
BB_filtered_image <- BB_filter(rgb_image = raw_image)
BB_fraction(filtered_image = BB_filtered_image)
}
\author{
Johanna Bloecher, \email{bloecher@fzp.czu.cz}
}
