% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB_vert_distribution.R
\name{BB_vert_distribution}
\alias{BB_vert_distribution}
\title{Vertical distribution of Brilliant Blue in Snow image}
\usage{
BB_vert_distribution(filtered_image, blue_selection = "all")
}
\arguments{
\item{filtered_image}{the returned matrix of the function BB_filter with values 0, 1, 2 and 3}

\item{blue_selection}{the blue to consider. By default, the value for blue_selection is 'all'
\itemize{
\item 'all' (all blues)
\item 'dark blue' (only dark blue),
\item 'medium blue' (only medium blues) or
\item 'light blue' (only light blues).
}}
}
\value{
Returns the vertical distribution of blue color as a vector.
}
\description{
A simple function that processes the result of BB_filter and returns a vector with fractions over the depth of an image of a sprinkling experiment with Brilliant Blue.
}
\examples{
BB_v_dist <- BB_vert_distribution(filtered_image = threshold_data, blue_selections = 'all')

}
\author{
Johanna Bloecher, \email{bloecher@fzp.czu.cz}
}
